refactor: use urfave/cli instead of yitsushi/go-commander
My CLI lib[^1] was a good exercise and an experiment[^2] 7 years ago. Back
then, I did not plan to use it that long.

I tried to keep the interface the same as it was before. Some changed,
but I don't think anyone will notice as only `dump` and `import`
changes.

* The `import` command now requires an `--input` flag instead of a
  positional argument.
* The `dump` command now required a `--yes-please` flag to make sure the
  user understands that the exported content will NOT be encrypted.
* The `dump` command now requires an `--output` flag, stdout is not the
  default behaviour. If someone really wants to dump secrets to
  stdout... Well, you can use `--output=/dev/stdout`.

Bug fixes:
* Import did not import Prefix on overwrite.

Closes #39

[^1]: https://github.com/yitsushi/go-commander
[^2]: https://github.com/yitsushi/totp-cli/issues/78#issuecomment-1639523393

Signed-off-by: Efertone <efertone@pm.me>