refactor: use Storage interface instead of direct struct
It can get quiet hard to test structs in Go and functions using them.
With this, the rest of the code can be mocked easier with a TestBackend
which is just sits in memory and no encryption. The encryption should be
tested on storage layer, but the rest of the application should care
only about the exposed functions.

related to #8 and #37

References:
- https://github.com/yitsushi/totp-cli/issues/8
- https://github.com/yitsushi/totp-cli/issues/37

Signed-off-by: Victoria Nadasdi <victoria@efertone.me>